Changes: -Rendering Reihenfolge gefixed. Der Cursor ist jetzt wieder über dem Inventar, hab einfach alle Screens einmal geflushed und dann halt den Cursor und Buttons danach zusammen -Save Data erstellt jetzt auf nötige überordner, falls diese nicht existieren, damit kann man dann in der Zukunft nen einfachen check machen ob Ordner und Dateien existieren für das Load Game/Profile Field, ob es halt existiert oder nicht -Item Texturen wurden nicht angezeigt, weil die Registry die alten Texture IDs gespeichert hat. Hab jetzt gemacht, dass während dem Load die Item TexturID und Größe auf die neue aus der Registry gesetzt wird -MaxHp und Hp werden im Spieler und Enemy jezt im Constructur endlich mal gesetzt, habe ich beim letzten mal vergessen -Die Position von der InventoryScreen Textur müsste jetzt je nach Screen Größe gefixed sein, war vorher bei Noah auf seinem 16:10 Screen verschoben.

+++ src/main/java/GUI/GUIInventorySlot.java
@@ -2,8 +2,8 @@ package GUI;
 
 public class GUIInventorySlot {
     //Position des Slots
-    public int PosX;
-    public int PosY;
+    public float PosX;
+    public float PosY;
     
     //Größe des Slots
     public int SlotSize;
@@ -11,7 +11,7 @@ public class GUIInventorySlot {
     //Index, also Position in der Liste
     public int SlotIndex;
 
-    public GUIInventorySlot(int PosX, int PosY, int SlotSize, int SlotIndex) { //Constructor
+    public GUIInventorySlot(float PosX, float PosY, int SlotSize, int SlotIndex) { //Constructor
         this.PosX = PosX;
         this.PosY = PosY;
         this.SlotSize = SlotSize;
